+++ app/Http/Controllers/ReviewController.php
@@ -68,12 +68,28 @@ class ReviewController extends Controller
      *     @OA\RequestBody(
      *          required=true,
      *          description="Pass data to add a new review",
-     *          @OA\JsonContent(type="object", ref="#/components/schemas/Review")
+     *          @OA\JsonContent(
+     *              @OA\Property(property="id", type="integer", readOnly=true, example=1),
+     *              @OA\Property(property="title", type="string", description="Review title", example="Gread place!!!"),
+     *              @OA\Property(property="rating", type="number", description="User rating", example=5),
+     *              @OA\Property(property="description", type="string", description="The main text of the review", example="Very cool restaurant, I liked everything"),
+     *              @OA\Property(property="like", type="integer", description="The number of likes under the review from other users", example=1, readOnly=true),
+     *              @OA\Property(property="place_id", type="integer", description="Id of the place for which the review is left", example=1),
+     *              @OA\Property(property="created_at", type="string", format="date-time", description="Initial creation timestamp", readOnly=true),
+     *              @OA\Property(property="updated_at", type="string", format="date-time", description="Last update timestamp", readOnly=true)
+     *          )
      *     ),
      *     @OA\Response(
      *          response=201,
      *          description="Success storing a new review",
      *          @OA\JsonContent(type="object", ref="#/components/schemas/Review")
+     *     ),
+     *     @OA\Response(
+     *          response=400,
+     *          description="Review already exists",
+     *          @OA\JsonContent(
+     *              @OA\Property(property="message", type="string", example="Review already exists")
+     *          )
      *     ),
      *     @OA\Response(
      *          response=401,
@@ -102,15 +118,22 @@ class ReviewController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
+        $validateReviewData = $request->validate([
             'title' => 'required|string',
             'rating' => 'required|numeric|min:1|max:5',
             'description' => 'required|string',
-            'place_id' => 'required',
-            'user_id' => 'required|unique:reviews'
+            'place_id' => 'required'
         ]);
 
-        $review = Review::create($request->all());
+        Place::findOrFail($request->get('place_id'));
+
+        if (Review::where('place_id', $request->get('place_id'))
+                ->where('user_id', auth()->user()->id)->first() !== null)
+            return response()->json(['message' => 'Review already exists'], 400);
+
+        $validateReviewData['user_id'] = auth()->user()->id;
+
+        $review = Review::create($validateReviewData);
 
         $this->placeRatingService->updatePlaceRatingAndReviewsCount($review);
 
@@ -124,7 +147,12 @@ class ReviewController extends Controller
      */
     public function show($id)
     {
-        return Review::findOrFail($id);
+        $review = Review::findOrFail($id);
+
+        if ($review->user_id != auth()->user()->id)
+            return response()->json(['message' => 'Access denied']);
+
+        return $review;
     }
 
     /**
@@ -172,13 +200,16 @@ class ReviewController extends Controller
      */
     public function update(Request $request, Review $review)
     {
-        $request->validate([
+        $validateReviewData = $request->validate([
             'title' => 'string',
             'rating' => 'numeric|min:1|max:5',
             'description' => 'string',
         ]);
 
-        $review->update($request->all());
+        if ($review->user_id !== auth()->user()->id)
+            return response()->json(['message' => 'Access denied'], 400);
+
+        $review->update($validateReviewData);
 
         $this->placeRatingService->updatePlaceRatingAndReviewsCount($review);
 
@@ -228,6 +259,10 @@ class ReviewController extends Controller
     public function destroy($id)
     {
         $review = Review::findOrFail($id);
+
+        if ($review->user_id !== auth()->user()->id)
+            return response()->json(['message' => 'Access denied'], 400);
+
         $review->comments()->delete();
         $review->delete();
